merged socket for BsimDma into socket already used by BsimCtrl
(the fd for a shared memory region is sent over with the first sglist call.
Portal number 666 is used to detect this special case on the receiving side (bsim)
and call the special magic to receive the fd from the socket.

BTW, atomicity of interactions on the socket is guaranteed with a mutex
on the bsim_exe side (the bsim side is synchronous wrt sockets anyway).