remove 'IMPORT_HOSTIF' support when instantiating user module in topgen.py
This support had been added in commit ed84be0f3f203a28697208b6d984d8311dd36546.
(If 'host' is needed as an extra parameter to a user module instantiation,
':host' can be added to the H2S_INTERFACES or S2H_INTERFACES line in the Makefile)

An example of this has been added to examples/echo.