Added byte write enables to PhysMemRequest and MemRequest
If BYTE_ENABLES is defined as a BSV preprocessor macro, byte write
enables are added to PhysMemRequest, in two fields: firstbe and
lastbe. The idea is to enable writing any number of contiguous bytes
even if it is less than the width of the shared memory data bus.

These have roughly the same semantics as in PCIE. The write enable in
firstbe apply to the first beat of a burst request and those in lastbe
apply to the last beat of a multi-beat burst request. Intervening beats
of a burst request enable the write of all beats of that burst.

For PCIe-based systems, only the first and last word of a write
request may have sub-word byte enables. There is still some work to do
in MemToPcie to adjust the starting address and burst length depending
on the write enables.