Ensure Makefile does not violate bsv path
Problem:

If --all was specified, Makefile rules were being generated for all
.bsv files even if there were multiple instances of the same
package/file name. Depending on the order in which make resolved
dependences, it could end up building a version of the package that
should have been shadowed in the BSV search path.

Solution:

Visit directories in order of the path, track which packages were
visited, and only emit dependences for the first instance found of any
package name.