Replace XDMA driver with latest Xilinx upstream
The AWS-provided drivers are old and have since-fixed bugs[1]. Whilst doing
this, clearly annotate every change made locally for Connectal so future
updates are easy to do.

Note that upstream now only maps the single BAR if XDMA_CONFIG_BAR_NUM is
defined, so we now map the BARs ourselves if they are not yet mapped.

[1] https://github.com/Xilinx/dma_ip_drivers/commit/d1f334b1372dbd7f2fa6f28a31b496779119f8d7